requirements.txt
torch
torchvision
cherry-rl
pandas
matplotlib
numpy